requirements.txt
discord.py
discord.py[voice]
PyNaCl
python-dotenv
rich
db-sqlite3
pillow
numpy